Move store to lib and split into files
