+++ frontend/src/Wrapper.js
@@ -7,7 +7,7 @@ import MomentUtils from '@date-io/moment'
 
 import App from './App'
 import client from './apollo/apolloClient'
-import { LoginStateProvider, MessagingStateProvider } from './store'
+import { LoginStateProvider, MessagingStateProvider } from './lib/store'
 import FocusOverlayProvider from './components/FocusOverlay'
 import InfoBoxProvider from './components/InfoBox'
 import InfoSnackbar from './components/InfoSnackbar'
