+++ frontend/src/components/BaseWorkspaceList.js
@@ -12,7 +12,7 @@ import {
 
 import { Privilege } from '../lib/permissions'
 import { exportWorkspace } from './WorkspaceNavBar'
-import { useMessageStateValue } from '../store'
+import { useMessageStateValue } from '../lib/store'
 import useRouter from '../lib/useRouter'
 import { useInfoBox } from './InfoBox'
