+++ frontend/src/components/ProjectNavBar.js
@@ -12,7 +12,7 @@ import {
 import { Privilege } from '../lib/permissions'
 import { PROJECT_BY_ID, PROJECTS_FOR_USER } from '../graphql/Query'
 import { DELETE_PROJECT } from '../graphql/Mutation'
-import { useMessageStateValue, useLoginStateValue } from '../store'
+import { useMessageStateValue, useLoginStateValue } from '../lib/store'
 import { useShareDialog } from '../dialogs/sharing'
 import useEditProjectDialog from '../dialogs/project/useEditProjectDialog'
 import useRouter from '../lib/useRouter'
