+++ frontend/src/components/WorkspaceNavBar.js
@@ -15,7 +15,7 @@ import client from '../apollo/apolloClient'
 import { EXPORT_QUERY, WORKSPACE_BY_ID, WORKSPACES_FOR_USER } from '../graphql/Query'
 import { DELETE_WORKSPACE } from '../graphql/Mutation'
 import useEditWorkspaceDialog from '../dialogs/workspace/useEditWorkspaceDialog'
-import { useMessageStateValue, useLoginStateValue } from '../store'
+import { useMessageStateValue, useLoginStateValue } from '../lib/store'
 import { useShareDialog } from '../dialogs/sharing'
 import { useInfoBox } from './InfoBox'
 import useRouter from '../lib/useRouter'
