+++ frontend/src/views/home/HomeView.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 import { Button, Grid, Typography, Container } from '@material-ui/core'
 
-import { useLoginStateValue } from '../../store'
+import { useLoginStateValue } from '../../lib/store'
 import UserViewContent from './UserViewContent'
 import useRouter from '../../lib/useRouter'
 import Auth from '../../lib/authentication'
