+++ frontend/src/views/mapper/ActiveCourse.js
@@ -9,7 +9,7 @@ import { Role } from '../../lib/permissions'
 import { Concept } from './concept'
 import { useCreateConceptDialog } from '../../dialogs/concept'
 import { useEditCourseDialog } from '../../dialogs/course'
-import { useLoginStateValue } from '../../store'
+import { useLoginStateValue } from '../../lib/store'
 import { useInfoBox } from '../../components/InfoBox'
 import DividerWithText from '../../components/DividerWithText'
 import useRouter from '../../lib/useRouter'
