+++ frontend/src/views/mapper/Course.js
@@ -6,7 +6,7 @@ import { useMutation } from 'react-apollo-hooks'
 
 import { Role } from '../../lib/permissions'
 import { Concept } from './concept'
-import { useLoginStateValue, useMessageStateValue } from '../../store'
+import { useLoginStateValue, useMessageStateValue } from '../../lib/store'
 import { useCreateConceptDialog } from '../../dialogs/concept'
 import { useEditCourseDialog } from '../../dialogs/course'
 import { UPDATE_COURSE_LINK } from '../../graphql/Mutation'
