+++ frontend/src/views/mapper/concept/Concept.js
@@ -13,7 +13,7 @@ import Tooltip from '@material-ui/core/Tooltip'
 import { Role } from '../../../lib/permissions'
 import { DELETE_CONCEPT, CREATE_CONCEPT_LINK } from '../../../graphql/Mutation'
 import cache from '../../../apollo/update'
-import { useMessageStateValue, useLoginStateValue } from '../../../store'
+import { useMessageStateValue, useLoginStateValue } from '../../../lib/store'
 import { useEditConceptDialog } from '../../../dialogs/concept'
 import { noPropagation } from '../../../lib/eventMiddleware'
 import generateTempId from '../../../lib/generateTempId'
