+++ frontend/src/views/user/UserView.js
@@ -3,7 +3,7 @@ import { makeStyles } from '@material-ui/core/styles'
 import { Button, CircularProgress, Typography } from '@material-ui/core'
 import { useMutation } from 'react-apollo-hooks'
 
-import { useLoginStateValue, useMessageStateValue } from '../../store'
+import { useLoginStateValue, useMessageStateValue } from '../../lib/store'
 import { Role } from '../../lib/permissions'
 import Auth from '../../lib/authentication'
 import useRouter from '../../lib/useRouter'
